Add weekly rescan cron job
The rescan happens every monday at 3:30am.

A rescan cron job is useful in case the contents of the /data directory
are changed by an entity other than NextCloud.  For instance, the user
may copy a file using `cp` into the /data directory and expect it to
show up in NextCloud eventually.